perf: defer dashboard items graph
